Moved functions into different files
