+++ assets/spritesheet.cfg
