+++ game/game.cpp
@@ -1,17 +1,98 @@
 // game.cpp : Defines the exported functions for the DLL application.
-//#define _DEBUG_WAYPOINTS_
+#define _DEBUG_WAYPOINTS_
 //#define _ENABLE_MIN_DELTA_TIME_
 
+// window
+#define SCREEN_WIDTH  226
+#define SCREEN_HEIGHT 248
+
+// layers
+#define BACKGROUND_LAYER 0
+#define DOT_LAYER 1
+#define PACMAN_LAYER 2
+#define GHOSTS_LAYER 3
+#define DEBUG_LAYER 99
+
+// binary directions
+#define BINARY_LEFT  0x8
+#define BINARY_RIGHT 0x4
+#define BINARY_UP    0x2
+#define BINARY_DOWN  0x1
+
+#define WAYPOINTS_LENGTH 65
+#define DOTS_LENGTH 246
+
+#define INVALID_WAYPOINT_INDEX -1
+
 #include <../include/ldk/ldk.h>
-#include "model.hpp"
-#include "controller.hpp"
+
+int Vec3ToIntDir(const ldk::Vec3);
+
+const ldk::Vec3 LEFT_DIR  = {-1,  0,  0};
+const ldk::Vec3 RIGHT_DIR = { 1,  0,  0};
+const ldk::Vec3 DOWN_DIR  = { 0, -1,  0};
+const ldk::Vec3 UP_DIR    = { 0,  1,  0};
+
+struct Entity
+{
+	ldk::Sprite sprite;
+	ldk::Vec3 direction;
+	ldk::Vec3 previousPosition;
+	float speed;
+};
+
+struct Waypoint
+{
+	ldk::Vec3 position;
+	int allowedDirections;
+
+	Waypoint()  { }
+
+	Waypoint(ldk::Vec3 pPos)
+	{
+		position = pPos;
+		allowedDirections = (BINARY_LEFT | BINARY_RIGHT | BINARY_UP | BINARY_DOWN);
+	}
+};
+
+struct Dot
+{
+	ldk::Vec3 position;
+	bool isSpecial;
+	bool isEnabled;
+
+	Dot() { }
+
+	Dot(ldk::Vec3 pPosition, bool pIsSpecial) :
+		isEnabled(true)
+	{
+		position = pPosition;
+		isSpecial = pIsSpecial;
+	}
+};
+
+struct GameState
+{
+	ldk::Material spritesheet;
+	ldk::Sprite background;
+	ldk::Sprite dotSprite;
+	int curWaypointIndex;
+	int playerPoints = 0;
+	ldk::Vec3 desiredDir;
+	float desiredAngle;
+	Entity pacman;
+} *gameState = nullptr;
+
+Waypoint allWaypoints[WAYPOINTS_LENGTH];
+Dot allDots[DOTS_LENGTH];
+
+#include "waypoint.cpp"
+#include "dot.cpp"
+#include "input.cpp"
 #include <iostream>
 
-void setupSprite();
-void InitializeWaypoints();
-void InitializeDots();
-void ReadInput(const float,Entity&);
-int InvertHorizontalDirection(const int);
+void ReadInput(const float, Entity&);
+void MovePacman(const float, Entity&);
 
 void gameInit(void* memory)
 {
@@ -20,8 +101,7 @@ void gameInit(void* memory)
 	gameState = (GameState*)memory;
 
 	// Background
-	gameState->spritesheet = ldk::render::loadMaterial("./assets/pacman.cfg");
-
+	gameState->spritesheet = ldk::render::loadMaterial("./assets/spritesheet.cfg");
 	gameState->background.position = { SCREEN_WIDTH*0.5, SCREEN_HEIGHT*0.5, BACKGROUND_LAYER };
 	gameState->background.color = { 1.0, 1.0, 1.0, 1.0 };
 	gameState->background.width = SCREEN_WIDTH;
@@ -37,8 +117,8 @@ void gameInit(void* memory)
 	gameState->dotSprite.angle = 0;
 	gameState->dotSprite.position = { 0, 0, DOT_LAYER };
 
-	InitializeWaypoints();
-	InitializeDots();
+	initializeWaypoints();
+	initializeDots();
     
 	// Pacman
 	gameState->pacman.sprite.color = { 1.0, 1.0, 1.0, 1.0 };
@@ -47,7 +127,7 @@ void gameInit(void* memory)
 	gameState->pacman.sprite.srcRect = { 473, 1, 13, 13 };
 	gameState->pacman.sprite.angle = 0;
 	gameState->pacman.sprite.position = { 114, 0, PACMAN_LAYER };
-	gameState->pacman.sprite.position.y = gameState->allWaypoints[WAYPOINTS_LENGTH-1].position.y;
+	gameState->pacman.sprite.position.y = allWaypoints[WAYPOINTS_LENGTH-1].position.y;
 	gameState->pacman.previousPosition = gameState->pacman.sprite.position;
 	gameState->pacman.speed = 60;
 	gameState->pacman.direction = { 0,0,0 };
@@ -62,7 +142,7 @@ void ResetGame()
 {
 	gameState->pacman.sprite.angle = 0;
 	gameState->pacman.sprite.position = { 114, 0, PACMAN_LAYER };
-	gameState->pacman.sprite.position.y = gameState->allWaypoints[WAYPOINTS_LENGTH-1].position.y;
+	gameState->pacman.sprite.position.y = allWaypoints[WAYPOINTS_LENGTH-1].position.y;
 	gameState->pacman.previousPosition = gameState->pacman.sprite.position;
 	gameState->pacman.speed = 60;
 	gameState->pacman.direction = { 0,0,0 };
@@ -72,331 +152,9 @@ void ResetGame()
 	gameState->curWaypointIndex = WAYPOINTS_LENGTH - 1;
 	gameState->playerPoints = 0;
     
-	Dot* allDotsPtr = gameState->allDots;
 	for (int i = 0; i < DOTS_LENGTH; i++)
 	{
-		allDotsPtr[i].isEnabled = true;
-	}
-}
-
-void InitializeWaypoints()
-{
-	Waypoint* allWaypointsPtr = gameState->allWaypoints;
-
-	// WAYPOINTS
-	// C1 - X 13
-	allWaypointsPtr[0] = Waypoint({ 13, SCREEN_HEIGHT - 12, 0 });
-	allWaypointsPtr[0].allowedDirections = (BINARY_DOWN | BINARY_RIGHT);
-	allWaypointsPtr[1] = Waypoint({ 13, SCREEN_HEIGHT - 43, 0 });
-	allWaypointsPtr[1].allowedDirections = (BINARY_DOWN | BINARY_RIGHT | BINARY_UP);
-	allWaypointsPtr[2] = Waypoint({ 13, SCREEN_HEIGHT - 67, 0 });
-	allWaypointsPtr[2].allowedDirections = (BINARY_RIGHT | BINARY_UP);
-	allWaypointsPtr[3] = Waypoint({ 13, SCREEN_HEIGHT - 163, 0 });
-	allWaypointsPtr[3].allowedDirections = (BINARY_RIGHT | BINARY_DOWN);
-	allWaypointsPtr[4] = Waypoint({ 13, SCREEN_HEIGHT - 187, 0 });
-	allWaypointsPtr[4].allowedDirections = (BINARY_RIGHT | BINARY_UP);
-	allWaypointsPtr[5] = Waypoint({ 13, SCREEN_HEIGHT - 211, 0 });
-	allWaypointsPtr[5].allowedDirections = (BINARY_RIGHT | BINARY_DOWN);
-	allWaypointsPtr[6] = Waypoint({ 13, SCREEN_HEIGHT - 235, 0 });
-	allWaypointsPtr[6].allowedDirections = (BINARY_RIGHT | BINARY_UP);
-
-	// C2 - X 29
-	allWaypointsPtr[7] = Waypoint({ 29, SCREEN_HEIGHT - 187, 0 });
-	allWaypointsPtr[7].allowedDirections = (BINARY_LEFT | BINARY_DOWN);
-	allWaypointsPtr[8] = Waypoint({ 29, SCREEN_HEIGHT - 211, 0 });
-	allWaypointsPtr[8].allowedDirections = (BINARY_RIGHT | BINARY_UP | BINARY_LEFT);
-
-	// C3 - X 53
-	allWaypointsPtr[9] = Waypoint({ 53, SCREEN_HEIGHT - 12, 0 });
-	allWaypointsPtr[9].allowedDirections = (BINARY_RIGHT | BINARY_LEFT | BINARY_DOWN);
-	allWaypointsPtr[10] = Waypoint({ 53, SCREEN_HEIGHT - 43, 0 });
-	allWaypointsPtr[10].allowedDirections = (BINARY_RIGHT | BINARY_LEFT | BINARY_DOWN | BINARY_UP);
-	allWaypointsPtr[11] = Waypoint({ 53, SCREEN_HEIGHT - 67, 0 });
-	allWaypointsPtr[11].allowedDirections = (BINARY_LEFT | BINARY_UP | BINARY_DOWN);
-	allWaypointsPtr[12] = Waypoint({ 53, SCREEN_HEIGHT - 115, 0 });
-	allWaypointsPtr[12].allowedDirections = (BINARY_LEFT | BINARY_RIGHT | BINARY_UP | BINARY_DOWN);
-	allWaypointsPtr[13] = Waypoint({ 53, SCREEN_HEIGHT - 163, 0 });
-	allWaypointsPtr[13].allowedDirections = (BINARY_LEFT | BINARY_RIGHT | BINARY_UP | BINARY_DOWN);
-	allWaypointsPtr[14] = Waypoint({ 53, SCREEN_HEIGHT - 187, 0 });
-	allWaypointsPtr[14].allowedDirections = (BINARY_RIGHT | BINARY_UP | BINARY_DOWN);
-	allWaypointsPtr[15] = Waypoint({ 53, SCREEN_HEIGHT - 211, 0 });
-	allWaypointsPtr[15].allowedDirections = (BINARY_LEFT | BINARY_UP);
-
-	// C4 - X 77
-	allWaypointsPtr[16] = Waypoint({ 77, SCREEN_HEIGHT - 43, 0 });
-	allWaypointsPtr[16].allowedDirections = (BINARY_DOWN | BINARY_RIGHT | BINARY_LEFT);
-	allWaypointsPtr[17] = Waypoint({ 77, SCREEN_HEIGHT - 67, 0 });
-	allWaypointsPtr[17].allowedDirections = (BINARY_UP | BINARY_RIGHT);
-	allWaypointsPtr[18] = Waypoint({ 77, SCREEN_HEIGHT - 92, 0 });
-	allWaypointsPtr[18].allowedDirections = (BINARY_DOWN | BINARY_RIGHT);
-	allWaypointsPtr[19] = Waypoint({ 77, SCREEN_HEIGHT - 115, 0 });
-	allWaypointsPtr[19].allowedDirections = (BINARY_DOWN | BINARY_LEFT | BINARY_UP);
-	allWaypointsPtr[20] = Waypoint({ 77, SCREEN_HEIGHT - 140, 0 });
-	allWaypointsPtr[20].allowedDirections = (BINARY_RIGHT | BINARY_DOWN | BINARY_UP);
-	allWaypointsPtr[21] = Waypoint({ 77, SCREEN_HEIGHT - 163, 0 });
-	allWaypointsPtr[21].allowedDirections = (BINARY_RIGHT | BINARY_LEFT | BINARY_UP);
-	allWaypointsPtr[22] = Waypoint({ 77, SCREEN_HEIGHT - 187, 0 });
-	allWaypointsPtr[22].allowedDirections = (BINARY_RIGHT | BINARY_LEFT | BINARY_DOWN);
-	allWaypointsPtr[23] = Waypoint({ 77, SCREEN_HEIGHT - 211, 0 });
-	allWaypointsPtr[23].allowedDirections = (BINARY_RIGHT | BINARY_UP);
-
-	// C5 - X 101
-	allWaypointsPtr[24] = Waypoint({ 101, SCREEN_HEIGHT - 12, 0 });
-	allWaypointsPtr[24].allowedDirections = (BINARY_LEFT | BINARY_DOWN);
-	allWaypointsPtr[25] = Waypoint({ 101, SCREEN_HEIGHT - 43, 0 });
-	allWaypointsPtr[25].allowedDirections = (BINARY_LEFT | BINARY_RIGHT | BINARY_UP);
-	allWaypointsPtr[26] = Waypoint({ 101, SCREEN_HEIGHT - 67, 0 });
-	allWaypointsPtr[26].allowedDirections = (BINARY_LEFT | BINARY_DOWN);
-	allWaypointsPtr[27] = Waypoint({ 101, SCREEN_HEIGHT - 92, 0 });
-	allWaypointsPtr[27].allowedDirections = (BINARY_LEFT | BINARY_RIGHT | BINARY_UP);
-	allWaypointsPtr[28] = Waypoint({ 101, SCREEN_HEIGHT - 163, 0 });
-	allWaypointsPtr[28].allowedDirections = (BINARY_LEFT | BINARY_DOWN);
-	allWaypointsPtr[29] = Waypoint({ 101, SCREEN_HEIGHT - 187, 0 });
-	allWaypointsPtr[29].allowedDirections = (BINARY_LEFT | BINARY_RIGHT | BINARY_UP);
-	allWaypointsPtr[30] = Waypoint({ 101, SCREEN_HEIGHT - 211, 0 });
-	allWaypointsPtr[30].allowedDirections = (BINARY_LEFT | BINARY_DOWN);
-	allWaypointsPtr[31] = Waypoint({ 101, SCREEN_HEIGHT - 235, 0 });
-	allWaypointsPtr[31].allowedDirections = (BINARY_LEFT | BINARY_RIGHT | BINARY_UP);
-
-	// Copy all information from the left side to the right side of the map
-	int halfallDotsLength = (int)(WAYPOINTS_LENGTH - 1) / 2;
-	for (int x = 0; x < halfallDotsLength; x++)
-	{
-		ldk::Vec3 curPos = allWaypointsPtr[x].position;
-		int realIndex = halfallDotsLength + x;
-		allWaypointsPtr[realIndex] = Waypoint({ SCREEN_WIDTH - curPos.x, curPos.y });
-		allWaypointsPtr[realIndex].allowedDirections = InvertHorizontalDirection(allWaypointsPtr[x].allowedDirections);
-	}
-    
-	allWaypointsPtr[WAYPOINTS_LENGTH-1] = 
-        Waypoint({ 113, allWaypointsPtr[WAYPOINTS_LENGTH-4].position.y, 0 });
-	allWaypointsPtr[WAYPOINTS_LENGTH-1].allowedDirections = (BINARY_LEFT | BINARY_RIGHT);
-}
-
-void InitializeDots()
-{
-	Waypoint* allWaypointsPtr = gameState->allWaypoints;
-	Dot* allDotsPtr = gameState->allDots;
-
-	ldk::Vec3 xOffset = { 7.9f, 0, 0 };
-	ldk::Vec3 yOffset = { 0, -7.9f, 0 };
-
-	int index = 0;
-    
-	// 1st column - 1st line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[0].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	// 1st column - 1st line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[0].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, true);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-
-	// 2nd column - 1st line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[9].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	// 2nd column - 1st line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[9].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-	// 3rd column - 1st line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[24].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-
-	// 1st column - 2 line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[1].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	// 1st column - 2 line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-
-	// 2nd column - 2nd line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[10].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	// 2nd column - 2nd line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[10].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-
-	// 3rd column - 2nd line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[16].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	// 3rd column - 2nd line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[16].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-
-	// 4th column - 2nd line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[25].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-
-
-	// 1st column - 3rd line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[2].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-
-
-	// 2nd column - 3rd line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[11].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-
-	// 3rd column - 3 line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[17].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-
-
-	// 2nd column - 4th line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[12].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset,false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-
-	// 1st column - 4th line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[3].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	// 1st column - 4th line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[3].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-
-	// 2nd column - 5th line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[13].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	// 2nd column - 5th line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[13].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-
-	// 3rd column - 6th line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[21].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-
-
-	// 4th column - 5th line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[28].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-
-	// 1st column - 5th line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[4].position, true);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-
-
-	// 2nd column - 5th line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[7].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-
-	// 3rd column - 6th line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[14].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	// 3rd column - 6th line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[14].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-
-	// 4th column - 7th line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[22].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	// 4th column - 7th line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[22].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-
-	// 1st column - 6th line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[5].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	// 1st column - 6th line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[5].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-
-	// 2nd column - 2nd line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[8].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-
-
-	// 4th column - 8th line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[23].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-
-
-	// 5th column - 7th line (vertical)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[30].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + yOffset, false);
-
-
-	// 1st column - 7th line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[6].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-
-
-	// 2nd column - 8th line (horizontal)
-	allDotsPtr[index++] = Dot(allWaypointsPtr[31].position, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-	allDotsPtr[index++] = Dot(allDotsPtr[index-1].position + xOffset, false);
-    
-    
-	// Copy all information from the left side to the right side of the map
-	int halfallDotsLength = (int)DOTS_LENGTH / 2;
-	for (int x = 0; x < halfallDotsLength; x++)
-	{
-		ldk::Vec3 curPos = allDotsPtr[x].position;
-		int realIndex = halfallDotsLength + x;
-		allDotsPtr[realIndex] = Dot({ SCREEN_WIDTH - curPos.x, curPos.y }, allDotsPtr[x].isSpecial);
+		allDots[i].isEnabled = true;
 	}
 }
 
@@ -412,7 +170,7 @@ void gameUpdate(float deltaTime)
     
 	if (ldk::input::isKeyDown(LDK_KEY_R))
 	{
-	ResetGame();
+		ResetGame();
 	}
 
 	ldk::render::spriteBatchBegin(gameState->spritesheet);
@@ -425,29 +183,25 @@ void gameUpdate(float deltaTime)
 
 	ldk::Sprite& dotSprite = gameState->dotSprite;
 
-	Dot* allDotsPtr = gameState->allDots;
-
 	// === DRAW ===
 	for (int x = 0; x < DOTS_LENGTH; x++)
 	{
-		if (allDotsPtr[x].isEnabled)
+		if (allDots[x].isEnabled)
 		{
-			dotSprite.position.x = allDotsPtr[x].position.x;
-			dotSprite.position.y = allDotsPtr[x].position.y;
-			dotSprite.width = dotSprite.height = (allDotsPtr[x].isSpecial) ? 8 : 2;
-			ldk::render::spriteBatchSubmit(gameState->dotSprite);
+			dotSprite.position.x = allDots[x].position.x;
+			dotSprite.position.y = allDots[x].position.y;
+			dotSprite.width = dotSprite.height = (allDots[x].isSpecial) ? 8 : 2;
+			ldk::render::spriteBatchSubmit(dotSprite);
 		}
 	}
 
 #ifdef _DEBUG_WAYPOINTS_ //_LDK_DEBUG_
-	Waypoint* allWaypoints = gameState->allWaypoints;
-
 	ldk::Vec3 defaultSize = {dotSprite.width, dotSprite.height};
 	ldk::Vec4 defaultColor = dotSprite.color;
 	int defaultLayer = dotSprite.position.z;
 	dotSprite.width = dotSprite.height = 4;
 	dotSprite.color = {0, 1, 0, 0.5};
-
+	
 	for (int x = 0; x < WAYPOINTS_LENGTH; x++)
 	{
 		dotSprite.position.x = allWaypoints[x].position.x;
@@ -455,7 +209,7 @@ void gameUpdate(float deltaTime)
 		dotSprite.position.z = DEBUG_LAYER;
 		ldk::render::spriteBatchSubmit(gameState->dotSprite);
 	}
-    
+    	
 	dotSprite.color = defaultColor;
 	dotSprite.width = defaultSize.x;
 	dotSprite.height = defaultSize.y;
@@ -466,80 +220,55 @@ void gameUpdate(float deltaTime)
 	ldk::render::spriteBatchEnd();
 }
 
-
-int InvertHorizontalDirection(const int allowedDirections)
+void MovePacman(const float deltaTime, Entity& pacman)
 {
-	int result = allowedDirections & (BINARY_UP | BINARY_DOWN);
+	pacman.sprite.position += pacman.direction * pacman.speed * deltaTime;
 
-	if ((allowedDirections & (BINARY_LEFT | BINARY_RIGHT)) != (BINARY_LEFT | BINARY_RIGHT))
+	if (pacman.direction.x != 0)
 	{
-		// Invert L and R
-		result |= ~allowedDirections & (BINARY_LEFT | BINARY_RIGHT);
+		// Pacman`s moving HORIZONTALLY
+		HandleHorizontalDots(pacman);
+		HandleHorizontalWaypoints(pacman);
+		
+		if (pacman.direction.x != 0)
+		{
+			gameState->curWaypointIndex = INVALID_WAYPOINT_INDEX;
+		}
 	}
-	else
+	else if (gameState->pacman.direction.y != 0)
 	{
-		result |= (BINARY_LEFT | BINARY_RIGHT);
+		// Pacman`s moving VERTICALLY
+		HandleVerticalDots(pacman);
+		HandleVerticalWaypoints(pacman);
+
+		if (pacman.direction.y != 0)
+		{
+			gameState->curWaypointIndex = INVALID_WAYPOINT_INDEX;
+		}
 	}
 
-	return result;
+	// Check bounds and teleport pacman, if needed
+	if (gameState->pacman.sprite.position.x < -5)
+	{
+		gameState->pacman.sprite.position.x = SCREEN_WIDTH + 4;
+	}
+	else if (gameState->pacman.sprite.position.x > SCREEN_WIDTH + 5)
+	{
+		gameState->pacman.sprite.position.x = -5;
+	}
+
+	pacman.previousPosition = pacman.sprite.position;
 }
 
-void ReadInput(const float deltaTime, Entity& pacman)
+int Vec3ToIntDir(const ldk::Vec3 direction)
 {
-	// Control
-	if (ldk::input::getKey(LDK_KEY_LEFT))
+	if (direction.x != 0)
 	{
-		gameState->desiredAngle = RADIAN(180);
-		gameState->desiredDir = LEFT_DIR;
+		return (direction.x > 0) ? BINARY_RIGHT : BINARY_LEFT;
 	}
-	else if (ldk::input::getKey(LDK_KEY_RIGHT))
+	if (direction.y != 0)
 	{
-		gameState->desiredAngle = 0;
-		gameState->desiredDir = RIGHT_DIR;
+		return (direction.y > 0) ? BINARY_UP : BINARY_DOWN;
 	}
-	else if (ldk::input::getKey(LDK_KEY_UP))
-	{
-		gameState->desiredAngle = RADIAN(90);
-		gameState->desiredDir = UP_DIR;
-	}
-	else if (ldk::input::getKey(LDK_KEY_DOWN))
-	{
-		gameState->desiredAngle = RADIAN(270);
-		gameState->desiredDir = DOWN_DIR;
-	}
-	else
-	{
-		gameState->desiredDir = ldk::Vec3::zero();
-	}
-
-	// Immediately direction change is only allowed on the same line of movement (Horizontally or Vertically)
-	if (!(gameState->desiredDir == ldk::Vec3::zero()))
-	{
-		// If he's walking
-		if (pacman.direction == ldk::Vec3::zero())
-		{
-			//If we`ve stopped on an Dot
-			if (gameState->curWaypointIndex != INVALID_WAYPOINT_INDEX)
-			{
-				// Check if this Dot allows us to move in the desired direction we want
-				Waypoint& waypoint = gameState->allWaypoints[gameState->curWaypointIndex];
-				int desiredDirAsInt = Vec3ToIntDir(gameState->desiredDir);
-				// Check if we can move in the desiredDir
-				if ((desiredDirAsInt & waypoint.allowedDirections) == desiredDirAsInt)
-				{
-					pacman.sprite.angle = gameState->desiredAngle;
-					pacman.direction = gameState->desiredDir;
-				}
-			}
-		}
-		else
-		{
-			// If the direction you wanna go is the inverse of the direction he is going
-			if (pacman.direction == (gameState->desiredDir * -1))
-			{
-				pacman.sprite.angle = gameState->desiredAngle;
-				pacman.direction = gameState->desiredDir;
-			}
-		}
-	}
+	return 0;
 }
